reorganized some converted files, upgraded deps.
