+++ package.json
@@ -43,7 +43,7 @@
     "rpgdicejs": "^2.0.0-rc.2",
     "sacjs": "^1.3.0",
     "socket.io": "^2.0.3",
-    "sqlite3": "^4.2.0",
+    "sqlite3": "^5.0.0",
     "trivial-logging": "^2.0.0",
     "trivialperms": "^1.0.0",
     "uuid": "^8.0.0"
@@ -90,7 +90,7 @@
     "sass": "^1.26.5",
     "socket.io-client": "^2.0.3",
     "ts-node": "^8.10.2",
-    "typescript": "^4.0.1-rc",
+    "typescript": "^4.0.3",
     "vue": "^2.6.11",
     "vue-codemirror": "^4.0.6",
     "vue-color": "^2.7.0",
