+++ server/api/resource-access/references.js
@@ -8,7 +8,7 @@ import _ from 'lodash';
 import * as dbMan from '../../managers/database';
 
 // Engines
-import suppEng from '../engines/supplement';
+import suppEng from '../../engines/supplement';
 
 // Utilities
 import { applyFilters } from '../../knex/utils';
